requirements.txt
transformers>=4.35.0
accelerate>=0.24.0
langchain>=0.1.0
langchain-community>=0.0.10
sentencepiece>=0.1.99
requests>=2.31.0
pillow>=10.0.0
timm>=0.9.0
duckduckgo-search>=4.1.0
huggingface-hub>=0.19.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
tqdm>=4.65.0 